+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0 - 2022-10-07]
+### Added
+- Substitute for Reservation class
+
+### Fixed
+- Reservation stream not generated properly
+
 ## [1.0.0 - 2022-09-29]
 ### Changed
 - First stable release